+++ app/src/main/java/woowacourse/shopping/cart/CartSelectionFragment.kt
@@ -1,60 +1,92 @@
 package woowacourse.shopping.cart
 
 import android.os.Bundle
+import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.lifecycle.ViewModelProvider
 import woowacourse.shopping.R
+import woowacourse.shopping.ShoppingApplication
+import woowacourse.shopping.cart.CartItem.PaginationButtonItem
+import woowacourse.shopping.databinding.FragmentCartSelectionBinding
+import woowacourse.shopping.product.catalog.ProductUiModel
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
-
-/**
- * A simple [Fragment] subclass.
- * Use the [CartSelectionFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
 class CartSelectionFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
+    private lateinit var binding: FragmentCartSelectionBinding
+    private val viewModel: CartViewModel by lazy {
+        ViewModelProvider(
+            this,
+            CartViewModelFactory(requireActivity().application as ShoppingApplication),
+        )[CartViewModel::class.java]
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
-        }
+        binding = FragmentCartSelectionBinding.inflate(layoutInflater)
+        binding.lifecycleOwner = this
+        setCartProductAdapter()
+        observeCartViewModel()
     }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Inflate the layout for this fragment
         return inflater.inflate(R.layout.fragment_cart_selection, container, false)
     }
 
-    companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment CartSelectionFragment.
-         */
-        // TODO: Rename and change types and number of parameters
-        @JvmStatic
-        fun newInstance(param1: String, param2: String) =
-            CartSelectionFragment().apply {
-                arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
+    private fun setCartProductAdapter() {
+        binding.recyclerViewCart.adapter =
+            CartAdapter(
+                cartItems = emptyList(),
+                onDeleteProductClick =
+                    DeleteProductClickListener { product ->
+                        viewModel.deleteCartProduct(CartItem.ProductItem(product))
+                    },
+                onPaginationButtonClick = viewModel::onPaginationButtonClick,
+                quantityControlListener = viewModel::updateQuantity,
+            )
+    }
+
+    private fun observeCartViewModel() {
+        viewModel.cartProducts.observe(this) { updateCartItems()
+        Log.d("test", "Cart Products: $it")
+        }
+        viewModel.isNextButtonEnabled.observe(this) { updateCartItems() }
+        viewModel.isPrevButtonEnabled.observe(this) { updateCartItems() }
+        viewModel.page.observe(this) { updateCartItems() }
+        viewModel.updatedItem.observe(this) { item ->
+            (binding.recyclerViewCart.adapter as CartAdapter).setCartItem(item)
+        }
+        viewModel.loadingState.observe(this) {
+            Log.d("LOADING_STATE", "State : $it")
+            when (it.isLoading) {
+                true -> {
+                    binding.shimmerFrameLayoutCartProducts.startShimmer()
+                    binding.shimmerFrameLayoutCartProducts.visibility = View.VISIBLE
+                }
+
+                false -> {
+                    Log.d("LOADING_STATE", "${it}")
+                    binding.shimmerFrameLayoutCartProducts.stopShimmer()
+                    binding.shimmerFrameLayoutCartProducts.visibility = View.GONE
                 }
             }
+        }
+    }
+
+    private fun updateCartItems() {
+        val products: List<ProductUiModel> = viewModel.cartProducts.value ?: return
+        val isNext: Boolean = viewModel.isNextButtonEnabled.value == true
+        val isPrev: Boolean = viewModel.isPrevButtonEnabled.value == true
+        val page: Int = viewModel.page.value ?: 0
+        val paginationItem = PaginationButtonItem(page + 1, isNext, isPrev)
+
+        val cartItems: List<CartItem> = products.map { CartItem.ProductItem(it) }
+        val cartItemsWithPaginationBtn =
+            if (cartItems.isEmpty()) cartItems else cartItems + paginationItem
+        (binding.recyclerViewCart.adapter as CartAdapter).setCartItems(cartItemsWithPaginationBtn)
     }
 }
